+++ apps/web/src/components/layout/components/v2/SampleWorkflowsBanner.tsx
@@ -1,0 +1,43 @@
+import { Group } from '@mantine/core';
+import { colors, useColorScheme } from '@novu/design-system';
+import { Button, Text } from '@novu/novui';
+import { useNavigate } from 'react-router-dom';
+
+export function SampleModeBanner() {
+  const { colorScheme } = useColorScheme();
+  const navigate = useNavigate();
+
+  return (
+    <div
+      style={{
+        width: '100%',
+        padding: 8,
+        backgroundColor: colorScheme === 'dark' ? '#13131A' : 'rgb(230 231 255)',
+        textAlign: 'center',
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        position: 'relative',
+      }}
+    >
+      <Group spacing={24}>
+        <Group spacing={8}>
+          <Text color={colors.black}>
+            You are currently viewing sample workflows. You can add your own by completing the Setup.
+          </Text>
+        </Group>
+        <Group spacing={20}>
+          <Button
+            onClick={() => {
+              navigate('/get-started');
+            }}
+            size="sm"
+            variant="transparent"
+          >
+            Complete Setup
+          </Button>
+        </Group>
+      </Group>
+    </div>
+  );
+}
